==> selection/scripts/06-psg.R <==
# ------------------------------------------------------------------------------------------------ #
# Hyphy/CodeML overlap
#
# This script processes the selection results generated by HyPhy and PAML. The results we work with
# include three separate selection testing runs:
#   - HyPhy: BUSTED-PH where Marine snakes are marked as 'test' (Terrestrial are 'background')
#   - HyPhy: BUSTED-PH where Terrestrial snakes are marked as 'test' (Marine are 'background')
#   - PAML: Branch-Site models were used to test for selection in Marine snakes (Foreground). To
#           ensure there was no signal of PS in the background branches, Marine snakes were dropped
#           and Site models were run on ONLY the Terrestrial snakes to test for positive selection.
#
# Positively selected genes (PSGs) are classified into three categories:
#   1. Marine specific: PSGs reported by both BUSTED-PH and PAML (Marine foreground/test).
#   2. Terrestrial specific: PSGs reported by both BUSTED-PH and PAML (Uses PAML drop-out tests for
#      evidence of selection in the Terrestrial snakes).
#   3. Shared: These are genes were there is evidence of positive selection in both the terrestrial
#      AND Marine snakes. These genes must be reported by both BUSTED-PH and PAML.

# ------------------------------------------------------------------------------------------------ #
# Libraries
suppressPackageStartupMessages({
  library(tidyverse)
  library(here)
  library(fs)

  # Loading for the pvalue correction functions I've written
  source(here('selection', 'scripts', 'hyphy-parsing', 'general.R'))
  source(here('selection', 'scripts', 'paml-parsing', 'general.R'))
})

# single-copy ortholog gene annotations
annotations <- read_csv(
  here('orthologs','ortholog-annotation','results','ortholog-annotation','orthologs.csv'),
  col_names = TRUE,
  col_types = cols()
) |>
  select(orthogroup, symbol)

# ------------------------------------------------------------------------------------------------ #
# Output directory
outdir <- here('selection','results','results-PSGs')
outdir.tables <- here('selection', 'results', 'results-tables')
dir_create(outdir)
dir_create(outdir.tables)

# ------------------------------------------------------------------------------------------------ #
# Load Hyphy and CodeML results (RDS files generated by 04-parse-hyphy.R)
bustedph <- read_rds(here('selection', 'r-data', 'busted-ph.rds'))
bustedph.terrestrial <- read_rds(here('selection', 'r-data', 'busted-ph-terrestrial.rds'))

# Keep only Branch-Site comparisons (remove M0 as this was simply used to get initial tree values)
paml <- read_csv(
  file = here('selection', 'results', 'paml', 'lrt-pval-comparison.csv'),
  col_names = TRUE,
  col_types = cols(),
  col_select = -signal
) |>
  filter(null_x == 'bsA1', alt_x == 'bsA', null_y == 'M1', alt_y == 'M2')

# ------------------------------------------------------------------------------------------------ #
# Correct P-values for multiple testing

# BUSTED-PH positively selected genes
bustedph$`test results corrected` <- pcorrBUSTEDPH(
  bustedph,
  p = 0.01,
  corrMethod = 'fdr'
) |>
  mutate(
    file = sub('.clean', '', file)
  ) |>
  left_join(annotations, by = join_by(file == orthogroup)) |>
  select(
    orthogroup = file,
    symbol,
    everything()
  )

write_csv(
  bustedph$`test results corrected`,
  file = here(outdir.tables,'bustedph-corrected.csv'),
  col_names = TRUE
)
write_rds(bustedph, file = here('selection', 'r-data', 'busted-ph.rds'), compress = 'gz')

# Positively selected genes (terrestrial)
bustedph.terrestrial$`test results corrected` <- pcorrBUSTEDPH(
  bustedph.terrestrial,
  p = 0.01,
  corrMethod = 'fdr'
) |>
  mutate(
    file = sub('.clean', '', file)
  ) |>
  left_join(annotations, by = join_by(file == orthogroup)) |>
  select(
    orthogroup = file,
    symbol,
    everything()
  )

write_csv(
  bustedph.terrestrial$`test results corrected`,
  file = here(outdir.tables,'bustedph-terrestrial-corrected.csv'),
  col_names = TRUE
)
write_rds(bustedph.terrestrial, here('selection', 'r-data', 'busted-ph-terrestrial.rds'), compress = 'gz')

# Positively selected genes (PAML): # 2,674
paml <- pcorrPAML(
  df = paml,
  p = 0.01,
  method = 'fdr'
) |>
  left_join(annotations, by = join_by(file == orthogroup)) |>
  select(orthogroup = file, symbol, everything())

paml |>
  write_csv(
    file = here(outdir.tables,'paml-corrected.csv'),
    col_names = TRUE
  )

# ------------------------------------------------------------------------------------------------ #
# Positively selected genes: Unique to Marine OR Terrestrial snakes

# Marine (Hydrophis) # 1,402
psg.marine.hyphy.paml <- bustedph$`test results corrected` |>
  left_join(paml) |>
  filter(
    str_detect(result, pattern = 'Selection associated with trait'),
    signal == 'PS_fg'
  ) |>
  pull(orthogroup)

write_lines(
  x = psg.marine.hyphy.paml,
  file = here(outdir,'PSGs-marine.txt')
)

# Terrestrial # 484
psg.terrestrial.hyphy.paml <- bustedph.terrestrial$`test results corrected` |>
  left_join(paml) |>
  filter(
    str_detect(result, pattern = 'Selection associated with trait'),
    signal == 'PS_bg'
  ) |>
  pull(orthogroup)

write_lines(
  x = psg.terrestrial.hyphy.paml,
  file = here(outdir, 'PSGs-terrestrial.txt')
)

# Sanity check: No Marine genes should be in Terrestrial genes
table(psg.marine.hyphy.paml %in% psg.terrestrial.hyphy.paml)

# ------------------------------------------------------------------------------------------------ #
# Positively selected genes: Signal in Marine AND Terrestrial

# PSGs shared by Marine and Terrestrial snakes # 532
psg.shared.hyphy.paml <- bustedph$`test results corrected`|>
  filter(str_detect(result, 'Selection acting on branches with phenotype')) |>
  select(orthogroup) |>
  left_join(bustedph.terrestrial$`test results corrected`) |>
  filter(str_detect(result, 'Selection acting on branches with phenotype')) |>
  select(orthogroup) |>
  left_join(paml) |>
  filter(signal == 'PS_fg_bg') |>
  pull(orthogroup)

write_lines(
  x = psg.shared.hyphy.paml,
  file = here(outdir, 'PSGs-shared.txt')
)

# Sanity check: No shared genes should be in Marine/Terrestrial unique PSGs
table(c(psg.marine.hyphy.paml, psg.terrestrial.hyphy.paml) %in% psg.shared.hyphy.paml)

# ------------------------------------------------------------------------------------------------ #
# Insignificant genes: Genes not under selection in either Marine OR Terrestrial OR Marine & Terrestrial

# All PSGs (unique/shared)
all.psg <- c(psg.marine.hyphy.paml, psg.terrestrial.hyphy.paml, psg.shared.hyphy.paml)

# Neutral/non-PSG genes in dataset # 6,236
paml$orthogroup[! paml$orthogroup %in% all.psg] |>
  write_lines(
    file = here(outdir, 'neutral-genes.txt')
  )
